+++ Avespoir.Core/Modules/Commands/ModeratorCommands/DBRoleList.cs
@@ -15,8 +15,8 @@ namespace Avespoir.Core.Modules.Commands {
 		public async Task DBRoleList(CommandObjects CommandObject) {
 			IMongoCollection<Roles> DBRolesCollection = MongoDBClient.Database.GetCollection<Roles>(typeof(Roles).Name);
 
-			FilterDefinition<Roles> DBRolesFilter = Builders<Roles>.Filter.Empty;
-			List<Roles> DBRolesList = await (await DBRolesCollection.FindAsync(DBRolesFilter).ConfigureAwait(false)).ToListAsync().ConfigureAwait(false);
+			FilterDefinition<Roles> DBGuildIDFilter = Builders<Roles>.Filter.Eq(Role => Role.GuildID, CommandObject.Guild.Id);
+			List<Roles> DBRolesList = await (await DBRolesCollection.FindAsync(DBGuildIDFilter).ConfigureAwait(false)).ToListAsync().ConfigureAwait(false);
 
 			List<object> DBRolesObjects = new List<object> { };
 			foreach (Roles DBRole in DBRolesList) {
@@ -26,8 +26,8 @@ namespace Avespoir.Core.Modules.Commands {
 			object[] DBRolesArray = DBRolesObjects.ToArray();
 			string DBRolesTableText = new TableFormatter().FormatObjects(DBRolesArray);
 
-			await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Message.Author.Mention + "DMをご確認ください！");
-			if (string.IsNullOrWhiteSpace(DBRolesTableText)) await CommandObject.Member.SendMessageAsync("何も登録されていません");
+			await CommandObject.Message.Channel.SendMessageAsync(string.Format(CommandObject.Language.DMMention, CommandObject.Message.Author.Mention));
+			if (string.IsNullOrWhiteSpace(DBRolesTableText)) await CommandObject.Member.SendMessageAsync(CommandObject.Language.ListNothing);
 			else await CommandObject.Member.SendMessageAsync(DBRolesTableText);
 		}
 	}