+++ Avespoir.Core/Modules/Commands/ModeratorCommands/DBUserAdd.cs
@@ -17,7 +17,7 @@ namespace Avespoir.Core.Modules.Commands {
 			try {
 				string[] msgs = CommandObject.CommandArgs.Remove(0);
 				if (msgs.Length == 0) {
-					await CommandObject.Message.Channel.SendMessageAsync("何も入力されていません");
+					await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.EmptyText);
 					return;
 				}
 
@@ -26,61 +26,68 @@ namespace Avespoir.Core.Modules.Commands {
 				uint msgs_RoleNum;
 
 				if (string.IsNullOrWhiteSpace(msgs[0])) {
-					await CommandObject.Message.Channel.SendMessageAsync("名前が空白またはNullです");
+					await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.EmptyName);
 					return;
 				}
 				msgs_Name = msgs[0];
 
 				if (string.IsNullOrWhiteSpace(msgs[1])) {
-					await CommandObject.Message.Channel.SendMessageAsync("IDが空白またはNullです");
+					await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.EmptyId);
 					return;
 				}
 				if (!ulong.TryParse(msgs[1], out msgs_ID)) {
-					await CommandObject.Message.Channel.SendMessageAsync("IDは数字でなければいけません");
+					await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.IdCouldntParse);
 					return;
 				}
 
 				if (string.IsNullOrWhiteSpace(msgs[2])) {
-					await CommandObject.Message.Channel.SendMessageAsync("RoleNumberが空白またはNullです");
+					await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.EmptyRoleNumber);
 					return;
 				}
 				if (!uint.TryParse(msgs[2], out msgs_RoleNum)) {
-					await CommandObject.Message.Channel.SendMessageAsync("RoleNumberは数字でなければいけません");
+					await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.RoleNumberNotNumber);
 					return;
 				}
 
 				IMongoCollection<AllowUsers> DBAllowUsersCollection = MongoDBClient.Database.GetCollection<AllowUsers>(typeof(AllowUsers).Name);
 				IMongoCollection<Roles> DBRolesCollection = MongoDBClient.Database.GetCollection<Roles>(typeof(Roles).Name);
+				FilterDefinition<AllowUsers> DBAllowUsersGuildIDFilter = Builders<AllowUsers>.Filter.Eq(AllowUser => AllowUser.GuildID, CommandObject.Guild.Id);
+				FilterDefinition<Roles> DBRolesGuildIDFilter = Builders<Roles>.Filter.Eq(Role => Role.GuildID, CommandObject.Guild.Id);
 
 				try {
 					FilterDefinition<AllowUsers> DBAllowUsersNameFilter = Builders<AllowUsers>.Filter.Eq(AllowUser => AllowUser.Name, msgs_Name);
-					AllowUsers DBAllowUsersName = await (await DBAllowUsersCollection.FindAsync(DBAllowUsersNameFilter).ConfigureAwait(false)).FirstAsync().ConfigureAwait(false);
+					FilterDefinition<AllowUsers> DBAllowUsersGuildIDNameFilter = Builders<AllowUsers>.Filter.And(DBAllowUsersGuildIDFilter, DBAllowUsersNameFilter);
+
+					AllowUsers DBAllowUsersName = await (await DBAllowUsersCollection.FindAsync(DBAllowUsersGuildIDNameFilter).ConfigureAwait(false)).FirstAsync().ConfigureAwait(false);
 					// if DBAllowUsersName is null, InvalidOperationException is a normal operation.
 
-					await CommandObject.Message.Channel.SendMessageAsync("その名前はすでに登録されています");
+					await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.NameRegisted);
 					return;
 				}
 				catch (InvalidOperationException) {
 					try {
 						FilterDefinition<AllowUsers> DBAllowUsersIDFilter = Builders<AllowUsers>.Filter.Eq(AllowUser => AllowUser.uuid, msgs_ID);
-						AllowUsers DBAllowUsersID = await (await DBAllowUsersCollection.FindAsync(DBAllowUsersIDFilter).ConfigureAwait(false)).FirstAsync().ConfigureAwait(false);
+						FilterDefinition<AllowUsers> DBAllowUsersGuildIDIDFilter = Builders<AllowUsers>.Filter.And(DBAllowUsersGuildIDFilter, DBAllowUsersIDFilter);
+						AllowUsers DBAllowUsersID = await (await DBAllowUsersCollection.FindAsync(DBAllowUsersGuildIDIDFilter).ConfigureAwait(false)).FirstAsync().ConfigureAwait(false);
 						// if DBRolesNum is null, InvalidOperationException is a normal operation.
 
-						await CommandObject.Message.Channel.SendMessageAsync("そのIDはすでに登録されています");
+						await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.IdRegisted);
 						return;
 					}
 					catch (InvalidOperationException) {
 						try {
 							FilterDefinition<Roles> DBRolesNumFilter = Builders<Roles>.Filter.Eq(Role => Role.RoleNum, msgs_RoleNum);
-							Roles DBRolesNum = await (await DBRolesCollection.FindAsync(DBRolesNumFilter).ConfigureAwait(false)).FirstAsync().ConfigureAwait(false);
+							FilterDefinition<Roles> DBRolesGuildIDNumFilter = Builders<Roles>.Filter.And(DBRolesGuildIDFilter, DBRolesNumFilter);
+							Roles DBRolesNum = await (await DBRolesCollection.FindAsync(DBRolesGuildIDNumFilter).ConfigureAwait(false)).FirstAsync().ConfigureAwait(false);
 							// if DBRolesNum is null, processes will not be executed from here.
 
 							if (!await Authentication.Confirmation(CommandObject)) {
-								await CommandObject.Channel.SendMessageAsync("認証に失敗しました、初めからやり直してください");
+								await CommandObject.Channel.SendMessageAsync(CommandObject.Language.AuthFailure);
 								return;
 							}
 
 							AllowUsers InsertAllowUserData = new AllowUsers {
+								GuildID = CommandObject.Guild.Id,
 								Name = msgs_Name,
 								uuid = msgs_ID,
 								RoleNum = msgs_RoleNum
@@ -88,18 +95,18 @@ namespace Avespoir.Core.Modules.Commands {
 							await DBAllowUsersCollection.InsertOneAsync(InsertAllowUserData).ConfigureAwait(false);
 
 							DiscordRole GuildRole = CommandObject.Guild.GetRole(DBRolesNum.uuid);
-							string ResultText = string.Format("Userデータベースに\n名前: {0}\nID: {1}\nRoleNumber: {2}({3})\nで登録しました！", InsertAllowUserData.Name, InsertAllowUserData.uuid, InsertAllowUserData.RoleNum, GuildRole.Name);
+							string ResultText = string.Format(CommandObject.Language.DBUserAddSuccess, InsertAllowUserData.Name, InsertAllowUserData.uuid, InsertAllowUserData.RoleNum, GuildRole.Name);
 							await CommandObject.Message.Channel.SendMessageAsync(ResultText);
 						}
 						catch (InvalidOperationException) {
-							await CommandObject.Message.Channel.SendMessageAsync("RoleNumberがRoleデータベースに存在しません");
+							await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.RoleNumberNotFound);
 							return;
 						}
 					}
 				}
 			}
 			catch (IndexOutOfRangeException) {
-				await CommandObject.Message.Channel.SendMessageAsync("必要箇所が入力されていません");
+				await CommandObject.Message.Channel.SendMessageAsync(CommandObject.Language.TypingMissed);
 			}
 		}
 	}