+++ Avespoir.Core/Modules/Events/ReadyEvent.cs
@@ -1,4 +1,4 @@
-﻿using Avespoir.Core.Configs;
+﻿using Avespoir.Core.JsonScheme;
 using Avespoir.Core.Database;
 using Avespoir.Core.Database.Schemas;
 using Avespoir.Core.Modules.Logger;
@@ -9,6 +9,7 @@ using MongoDB.Driver;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Avespoir.Core.Configs;
 
 namespace Avespoir.Core.Modules.Events {
 
@@ -23,30 +24,22 @@ namespace Avespoir.Core.Modules.Events {
 			BaseDiscordClient ClientBase = ReadyEventObjects.Client;
 			Client.Bot.UpdateStatusAsync(StartingStatus, UserStatus.DoNotDisturb).ConfigureAwait(false);
 
+			
+
 			Log.Info($"{ClientBase.CurrentUser.Username} Bot Ready!");
 
-			DBCount().ConfigureAwait(false);
+			StartStatus().ConfigureAwait(false);
 			Log.Debug("ReadyEvent " + "End...");
 			return Task.CompletedTask;
 		}
 
-		static async Task DBCount() {
+		static async Task StartStatus() {
 			while (true) {
-				try {
-					IMongoCollection<AllowUsers> DBAllowUsersCollection = MongoDBClient.Database.GetCollection<AllowUsers>(typeof(AllowUsers).Name);
-					List<AllowUsers> DBAllowUsersAllList = await (await DBAllowUsersCollection.FindAsync(_ => true).ConfigureAwait(false)).ToListAsync().ConfigureAwait(false);
-
-					DiscordGame ReadyStatus = new DiscordGame() {
-						Name = CommandConfig.PublicPrefix + "help" + " | " + DBAllowUsersAllList.Count + " " + "Users!",
-					};
-					await Client.Bot.UpdateStatusAsync(ReadyStatus, UserStatus.Online).ConfigureAwait(false);
-				}
-				catch (InvalidOperationException) {
-					DiscordGame ReadyStatus = new DiscordGame() {
-						Name = CommandConfig.PublicPrefix + "help",
-					};
-					await Client.Bot.UpdateStatusAsync(ReadyStatus, UserStatus.Online).ConfigureAwait(false);
-				}
+				DiscordGame ReadyStatus = new DiscordGame() {
+					Name = CommandConfig.PublicPrefix + "help",
+				};
+				
+				await Client.Bot.UpdateStatusAsync(ReadyStatus, UserStatus.Online).ConfigureAwait(false);
 
 				await Task.Delay(1000).ConfigureAwait(false);
 			}