+++ CHANGELOG.md
@@ -7,7 +7,43 @@ If it changes significantly or very much, the first digit will move.
 
 ***
 
-## [Beta 2.1](https://github.com/Avespoir-Project/Avespoir/tree/Beta-2.1) - 2020-07-09
+## [Beta 3.0](https://github.com/Fairy-Phy/Avespoir/tree/Beta-3.0) - 2020-08-31
+
+### Added
+
+* Database Schemes
+  * AllowUsers and Roles: GuildID
+  * GuildConfig
+
+* Multi Language Support (This version ja-JP only)
+
+* Server Configs
+  * Whitelist Enable/Disable
+  * LeaveBan Enable/Disable
+  * Custom Public Prefix
+  * Custom Moderator Prefix
+  * LogChannel
+  * Language (This version ja-JP only)
+
+* Public Command: Roll (Long type random)
+
+### Changed
+
+* Repository Change: Organizations -> Personal
+
+* Level System: Exp to First Sender -> First Sender and End Sender
+
+* Database Process: Target Proparty -> GuildID and Target Proparty
+
+* Database Scheme: LogChannel -> GuildConfig
+
+* Level System DatabaseMethods -> Unified Database DatabaseMethods
+
+### Removed
+
+* Lunetrip Project(Experimental) (so hard system...)
+
+## [Beta 2.1](https://github.com/Fairy-Phy/Avespoir/tree/Beta-2.1) - 2020-07-09
 
 ### Fixed
 
@@ -26,7 +62,7 @@ If it changes significantly or very much, the first digit will move.
 * Level System Calculation(ExpScale 1.0x -> 0.5x)
 * Level System Check NextMessage System -> Message Stack System
 
-## [Beta 2.0](https://github.com/Avespoir-Project/Avespoir/tree/Beta-2.0) - 2020-05-19
+## [Beta 2.0](https://github.com/Fairy-Phy/Avespoir/tree/Beta-2.0) - 2020-05-19
 
 ### Fixed
 
@@ -45,7 +81,7 @@ If it changes significantly or very much, the first digit will move.
 
 * self made logger -> log4net
 
-## [Beta 1.2](https://github.com/Avespoir-Project/Avespoir/tree/Beta-1.2) - 2020-05-19
+## [Beta 1.2](https://github.com/Fairy-Phy/Avespoir/tree/Beta-1.2) - 2020-05-19
 
 ### Added
 
@@ -59,13 +95,13 @@ If it changes significantly or very much, the first digit will move.
 
 * Botowner Command: DBChannelAdd
 
-## [Beta 1.1.1](https://github.com/Avespoir-Project/Avespoir/tree/Beta-1.1.1) - 2020-03-26
+## [Beta 1.1.1](https://github.com/Fairy-Phy/Avespoir/tree/Beta-1.1.1) - 2020-03-26
 
 ### Fixed
 
 * Moderator Command Allow Botowner
 
-## [Beta 1.1](https://github.com/Avespoir-Project/Avespoir/tree/Beta-1.1) - 2020-03-25
+## [Beta 1.1](https://github.com/Fairy-Phy/Avespoir/tree/Beta-1.1) - 2020-03-25
 
 ### Added
 
@@ -74,7 +110,7 @@ If it changes significantly or very much, the first digit will move.
   * DBUserChangeRole Command
 * Bot status display allow users count
 
-## [Beta 1.0](https://github.com/Avespoir-Project/Avespoir/tree/Beta-1.0) - 2020-03-24
+## [Beta 1.0](https://github.com/Fairy-Phy/Avespoir/tree/Beta-1.0) - 2020-03-24
 
 Start Beta version!
 
@@ -86,7 +122,7 @@ Start Beta version!
 
 * Test Code -> Production specification
 
-## [Alpha 4.0 (Final Update)](https://github.com/Avespoir-Project/Avespoir/tree/Alpha-4.0) - 2020-03-17
+## [Alpha 4.0 (Final Update)](https://github.com/Fairy-Phy/Avespoir/tree/Alpha-4.0) - 2020-03-17
 
 Alpha version final update.
 
@@ -129,7 +165,7 @@ Alpha version final update.
 * Client.Main args
 * Test command(send check)
 
-## [Alpha 3.0](https://github.com/Avespoir-Project/Avespoir/tree/Alpha-3.0) - 2020-03-11
+## [Alpha 3.0](https://github.com/Fairy-Phy/Avespoir/tree/Alpha-3.0) - 2020-03-11
 
 ### Added
 
@@ -149,7 +185,7 @@ Alpha version final update.
 
 * DSharpPlus.CommandNext
 
-## [Alpha 2.2(3.2.2 commit)](https://github.com/Avespoir-Project/Avespoir/tree/Alpha-2.2) - 2020-03-03
+## [Alpha 2.2(3.2.2 commit)](https://github.com/Fairy-Phy/Avespoir/tree/Alpha-2.2) - 2020-03-03
 
 The version name in the commit is incorrect.
 
@@ -168,7 +204,7 @@ The version name in the commit is incorrect.
 * Message.MessageEvent -> Events.MessageEvent
 * Avoid deadlock(?)
 
-## [Alpha 2.0](https://github.com/Avespoir-Project/Avespoir/tree/Alpha-2.0) - 2020-03-02
+## [Alpha 2.0](https://github.com/Fairy-Phy/Avespoir/tree/Alpha-2.0) - 2020-03-02
 
 ### Added
 
@@ -196,7 +232,7 @@ The version name in the commit is incorrect.
 
 * DSharpPlus.CommandNextModule RegisterCommands
 
-## [Alpha 1.1](https://github.com/Avespoir-Project/Avespoir/tree/Alpha-1.1) - 2020-02-23
+## [Alpha 1.1](https://github.com/Fairy-Phy/Avespoir/tree/Alpha-1.1) - 2020-02-23
 
 ### Added
 
@@ -211,7 +247,7 @@ The version name in the commit is incorrect.
 * CommandModuleNext method move CommandRegister class
 * MessageCreated -> MessageEvent
 
-## [Alpha 1.0](https://github.com/Avespoir-Project/Avespoir/tree/Alpha-1.0) - 2020-02-21
+## [Alpha 1.0](https://github.com/Fairy-Phy/Avespoir/tree/Alpha-1.0) - 2020-02-21
 
 ### Added